requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
line-bot-sdk==3.11.0
anthropic==0.39.0
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
gitpython==3.1.43
jinja2==3.1.4
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
